+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  * Fix Lua `find_home("all")` axis order bug.
  * Add `axis_order` to `move`.
  * Add `tool.seeder_tip_z_offset`.
+ * Add `get_weeds` and `get_plants` Lua helpers. Add point support to `sort` Lua helper.
 
 # 15.4.11